requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
asyncio==3.4.3
python-multipart==0.0.6
GitPython==3.1.40
aiofiles==23.2.1
